requirements.txt
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
pystray>=0.19.0
Pillow>=10.0.0
PyYAML>=6.0.0
lxml>=4.9.0
customtkinter>=5.2.0
